backend/worker/requirements.txt
requests==2.31.0
kafka-python==2.0.2
redis==5.0.1
protobuf==4.25.1
gtfs-realtime-bindings==1.0.0

